requirements.txt
glfw==2.9.0
imgui==2.0.0
numpy==1.24.1
Pillow==11.2.1
pyglm==2.8.2
PyOpenGL==3.1.6
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
trimesh==3.21.5
trimesh==4.6.10
